[SGMR-377] RDP 알고리즘을 적용한 러닝 저장 API 개선 (#54)
* feat: 러닝 및 코스 테이블 데이터 URL 필드 추가

* feat: Running, Course 테이블 수정사항 기존 코드에 반영

* feat: 시계열을 Multipart 객체로 받아서 위경도 추출 + 속도 계산 + 상대시간 변환하는 로직 구현

* feat: 시계열을 Multipart 객체로 받아서 위경도 추출 + 속도 계산 + 상대시간 변환하는 로직 구현

* feat: 위경도 객체에 RDP 알고리즘 적용

* feat: Raw 시계열, RDP 알고리즘을 적용한 시계열, 러닝 스크린샷을 S3에 저장하는 로직 구현

* feat: RunningServiceMapper 테스트 코드 추가

* feat: RDP, 평균 고도 계산 로직을 적용한 러닝/코스 저장 API 고도화

* feat: MulitpartFile이 Null이거나 Empty할 때 검증 로직 추가

* chore: 러닝 테이블명 기존 running_record로 변경

* chore: Saved 필드 제거, screenShot 용어 통일

* feat: .jsonl 파일의 형식이 잘못됐을 때에 대한 테스트 추가

* chore: S3에 업로드할 때 로그 수정

* feat: 스크린샷 데이터가 Null이지만 비어있는 케이스에 대한 테스트 추가

* fix: 기존 러닝 저장 API에서 시계열을 함께 요청하는 로직을 제거

+++ src/main/java/soma/ghostrunner/domain/running/api/dto/request/RunRecordRequest.java
@@ -15,10 +15,10 @@ public class RunRecordRequest {
     private Double distance;    // km
 
     @NotNull @PositiveOrZero
-    private Integer elevationGain;      // m, 고도
+    private Double elevationGain;      // m, 고도
 
     @NotNull @NegativeOrZero
-    private Integer elevationLoss;
+    private Double elevationLoss;
 
     @NotNull @PositiveOrZero
     private Long duration;      // 초 단위
